+++ packages/plugin-packages/mashroom-security-provider-openid-connect/type-definitions/index.ts
@@ -11,6 +11,7 @@ export type Mode = 'OIDC' | 'OAuth2';
 export type ClientConfiguration = {
     issuerDiscoveryUrl: string | undefined | null;
     issuerMetadata: IssuerMetadata | undefined | null;
+    rejectUnauthorized: boolean;
     clientId: string;
     clientSecret: string;
     redirectUrl: string;